+++ ledger/ledger-on-sql/src/main/scala/com/daml/ledger/on/sql/queries/ReadQueries.scala
@@ -4,7 +4,7 @@
 package com.daml.ledger.on.sql.queries
 
 import com.daml.ledger.on.sql.Index
-import com.daml.ledger.participant.state.kvutils.api.LedgerEntry
+import com.daml.ledger.participant.state.kvutils.api.LedgerRecord
 import com.daml.ledger.validator.LedgerStateOperations.{Key, Value}
 
 import scala.collection.immutable
@@ -13,7 +13,7 @@ import scala.util.Try
 trait ReadQueries {
   def selectLatestLogEntryId(): Try[Option[Index]]
 
-  def selectFromLog(start: Index, end: Index): Try[immutable.Seq[(Index, LedgerEntry)]]
+  def selectFromLog(start: Index, end: Index): Try[immutable.Seq[(Index, LedgerRecord)]]
 
   def selectStateValuesByKeys(keys: Seq[Key]): Try[immutable.Seq[Option[Value]]]
 }
