+++ ledger/participant-state/kvutils/src/main/scala/com/daml/ledger/participant/state/kvutils/api/LedgerReader.scala
@@ -23,13 +23,13 @@ trait LedgerReader extends ReportsHealth {
     *
     * In case an offset is not specified, all updates must be streamed from the oldest known state.
     * Each update is defined either as an opaque log entry ID and an
-    * envelope ([[com.daml.ledger.participant.state.kvutils.api.LedgerEntry.LedgerRecord]]).
+    * envelope ([[com.daml.ledger.participant.state.kvutils.api.LedgerRecord]]).
     *
     * @param startExclusive offset right after which updates must be streamed; in case not specified updates
     *                       must be returned from the beginning
     * @return stream of updates
     */
-  def events(startExclusive: Option[Offset]): Source[LedgerEntry, NotUsed]
+  def events(startExclusive: Option[Offset]): Source[LedgerRecord, NotUsed]
 
   /**
     * Get the ledger's ID from which this reader instance streams events.
@@ -45,7 +45,7 @@ object LedgerReader {
   /**
     * Default initial ledger configuration used by [[KeyValueParticipantStateReader]].
     */
-  val DefaultConfiguration = Configuration(
+  val DefaultConfiguration: Configuration = Configuration(
     generation = 0,
     timeModel = TimeModel.reasonableDefault,
     maxDeduplicationTime = Duration.ofDays(1),
